quering and mutation of tombstone fixed
