+++ tombbuilder/src/components/appEditor/editor/reactCodeEditor/hooks/useLoadTombstone.js
@@ -1,0 +1,10 @@
+import {useMutation,useQuery} from "react-query";
+import axios from "axios";
+import { useState,useEffect } from "react";
+
+const fetchTombstone=()=>{
+    return axios.get("http://localhost:4000/tombstone")
+    }
+export const useLoadTombstone=()=>{
+      return useQuery('tombstone',fetchTombstone);
+}
