+++ tombbuilder/src/components/appEditor/editor/reactCodeEditor/hooks/useSaveTombstone.js
@@ -1,0 +1,10 @@
+import {useMutation,useQuery} from "react-query";
+import axios from "axios";
+import { useState,useEffect } from "react";
+
+const addTombstone=(tombstone)=>{
+    return axios.post("http://localhost:4000/tombstone",tombstone);
+  }
+ export const useSaveTombstone=()=>{
+    return useMutation(addTombstone);
+  }
